Reconcile some typedefs with Musl's versions.

+++ basics/include/__typedef_blksize_t.h
@@ -1,7 +1,6 @@
 #ifndef __wasm_basics___typedef_blksize_t_h
 #define __wasm_basics___typedef_blksize_t_h
 
-/* Define these as 64-bit signed integers to support files larger than 2 GiB. */
-typedef long long blksize_t;
+typedef long blksize_t;
 
 #endif
